requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==18.2.0
beautifulsoup4==4.7.1
billiard==3.5.0.5
certifi==2018.11.29
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
Jinja2==2.10
kafka-python==1.4.3
lxml==4.3.0
MarkupSafe==1.1.0
multidict==4.5.2
pkg-resources==0.0.0
pytz==2018.9
redis==3.1.0
requests==2.21.0
soupsieve==1.6.2
typing-extensions==3.7.2
urllib3==1.24.1
uWSGI==2.0.17.1
vine==1.2.0
web.py==0.40.dev1
yarl==1.3.0
